DIRECTMEMORY-42 : OffHeapMemoryBuffer store, allocate and free function's parameters consistency
git-svn-id: https://svn.apache.org/repos/asf/incubator/directmemory/trunk@1227843 13f79535-47bb-0310-9956-ffa450edef68

+++ directmemory-cache/src/main/java/org/apache/directmemory/memory/MemoryManagerService.java
@@ -50,6 +50,6 @@ public interface MemoryManagerService
 
     OffHeapMemoryBuffer getActiveBuffer();
 
-    Pointer allocate( int size, int expiresIn, int expires );
+    Pointer allocate( int size, long expiresIn, long expires );
 
 }
